Add mixin for Scrapy RDBMS pipelines
