+++ src/scraper_carrefour/pipeline_rdbms.py
@@ -5,35 +5,21 @@ from itemadapter import ItemAdapter
 from scrapy import Item
 from scrapy.exceptions import DropItem
 from scrapy.spiders import Spider
-from sqlalchemy import create_engine, select
-from sqlalchemy.orm import Session
+from sqlalchemy import select
 
-from models.base import Base
 from models.category import Category
 from models.price import Price
 from models.product import Product
 from models.source import Origin, Source
-from utils.database import DEFAULT_DATABASE_URL
+from utils.database import RDBMSPipelineMixin
 
 from .items import ProductItem
 
 
-class ProductPipeline:
+class ProductPipeline(RDBMSPipelineMixin):
     """
     Scrapy pipeline used to store items into a RDBMS via SQLAlchemy.
     """
-
-    def open_spider(self, spider: Spider):
-        database_url = spider.settings.get("DATABASE_URL", DEFAULT_DATABASE_URL)
-
-        engine = create_engine(database_url)
-        Base.metadata.create_all(engine)
-
-        self.db_engine = engine
-        self.db_session = Session(self.db_engine)
-
-    def close_spider(self, spider: Spider):
-        self.db_session.close()
 
     def process_item(self, item: Item, spider: Spider):
         logger = getLogger(__name__)
